backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
websockets>=11.0
httpx>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
